Add metrics to the entity view.
